+++ packages/plugin-gallery/web/src/gallery-viewer.jsx
@@ -274,6 +274,7 @@ class GalleryViewer extends React.Component {
       >
         {size?.width ? (
           <ProGallery
+            options={{ allowContextMenu: true }}
             domId={this.domId}
             allowSSR={!!seoMode}
             items={items}